+++ NEWS.md
@@ -5,6 +5,7 @@
 - Add grid_compat_tol argument to exact_extract and relax from previously
   hardcoded default
 - Avoid error when include_area = TRUE and polygon does not intersect raster
+- Support terra SpatRaster inputs to coverage_fraction
 
 # version 0.7.2